Get group memberships from Azure
The groups include transitive memberships. The first request to the
Microsoft Graph API returns only the guids of the groups, therefore one
additional request has to be made for each group to get the name. To
make sure this doesn't slow down the login process for users with many
group memberships, the requests are run in parallel by using Guzzle
promises, because that's the library that the oauth2-client library by
league uses. oauth2-client does not provide parallel requests though. It
would be better if this logic would be implemented in the lower level
libraries, but it works for now.

Although it's already included as an indirect dependency, this adds
guzzlehttp/promises as a direct dependency, because it's directly used.